+++ cognite/powerops/client/_generated/cogshop1/__init__.py
@@ -1,3 +1,0 @@
-from cognite.powerops.client._generated.cogshop1._api_client import CogShop1Client
-
-__all__ = ["CogShop1Client"]