+++ cognite/powerops/client/powerops_client.py
@@ -12,14 +12,9 @@ from cognite.client import CogniteClient, global_config
 from cognite.powerops.client.shop.cogshop_api import CogShopAPI
 from cognite.pygen.utils.external_id_factories import ExternalIdFactory
 
-from ._generated.afrr_bid import AFRRBidAPI
-from ._generated.assets import PowerAssetAPI
-from ._generated.cogshop1 import CogShop1Client
-from ._generated.day_ahead_bid import DayAheadBidAPI
 from ._generated.v1 import PowerOpsModelsV1Client
 from ._generated.v1.data_classes._core import DomainModelWrite
 from .data_set_api import DataSetsAPI
-from .shop.dayahead_trigger_api import DayaheadTriggerAPI
 from .shop.shop_run_api import SHOPRunAPI
 
 # max_domain = max_total (255) - uuid (32) + separator (1)  noqa: ERA001
@@ -38,13 +33,7 @@ class PowerOpsClient:
     ):
         self.cdf = client
         self.datasets = DataSetsAPI(self.cdf, read_dataset, write_dataset, monitor_dataset)
-        self.cog_shop1 = CogShop1Client(self.cdf)
-        self.assets = PowerAssetAPI(self.cdf)
-        self.afrr_bid = AFRRBidAPI(self.cdf)
-        self.day_ahead_bid = DayAheadBidAPI(self.cdf)
         self.shop = SHOPRunAPI(self.cdf, self.datasets.write_dataset_id)
-        self.workflow = DayaheadTriggerAPI(self.cdf, self.datasets.write_dataset_id)
-
         self.v1 = PowerOpsModelsV1Client(self.cdf)
         self.cogshop = CogShopAPI(self.cdf, self.v1)