+++ cognite/powerops/client/shop/__init__.py
@@ -1,7 +1,3 @@
-from .dayahead_trigger_api import DayaheadTriggerAPI
 from .shop_run_api import SHOPRunAPI
 
-__all__ = [
-    "DayaheadTriggerAPI",
-    "SHOPRunAPI",
-]
+__all__ = ["SHOPRunAPI"]